% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Absenteeism.R
\docType{data}
\name{Absenteeism}
\alias{Absenteeism}
\title{Absenteeism}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 146 rows and 5 columns.
}
\usage{
Absenteeism
}
\description{
Absenteeism dataset
}
\keyword{datasets}
