% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChildHealth.R
\docType{data}
\name{ChildHealth}
\alias{ChildHealth}
\title{ChildHealth}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1236 rows and 8 columns.
}
\usage{
ChildHealth
}
\description{
Child Health (babies dataset from OpenIntro)
}
\keyword{datasets}
