% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HolzingerSwineford1939.R
\docType{data}
\name{HolzingerSwineford1939}
\alias{HolzingerSwineford1939}
\title{HolzingerSwineford1939}
\format{
An object of something
}
\usage{
data(HolzingerSwineford1939)
}
\description{
Anscombe dataset
}
\examples{
data(HolzingerSwineford1939)

}
\keyword{datasets}
