% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotorTrends.R
\docType{data}
\name{MotorTrends}
\alias{MotorTrends}
\title{MotorTrends}
\format{
An object of something
}
\usage{
data(MotorTrends)
}
\description{
Anscombe dataset
}
\examples{
data(MotorTrends)

}
\keyword{datasets}
