% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Personality.R
\docType{data}
\name{Personality}
\alias{Personality}
\title{Personality}
\format{
An object of class \code{data.frame} with 2800 rows and 28 columns.
}
\usage{
Personality
}
\description{
Big Five Personality Inventory dataset
}
\keyword{datasets}
